Extract NavMenu component from NavBar.
